requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==1.17.1
channels==4.3.1
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.6
daphne==4.2.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
MT5Manager==5.0.5200
numpy==2.3.2
psycopg==3.2.9
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyodbc==5.2.0
pyOpenSSL==25.1.0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
zope.interface==7.2
